pyproject.toml
[tool.poetry]
name = "ember-django"
version = "0.1.0"
description = "Ember API with Django and DRF"
authors = ["Rabscuttler <laurence.watson@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.17.2"
gunicorn = "^20.1.0"
django-csvimport = "^2.16"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
